Fix identicon, implement last used provider login
