+++ src/containers/App/index.js
@@ -17,8 +17,8 @@ import { getHtmlConfig, isFeatureEnabled } from 'utils/features'
 import HeaderContainer from 'containers/HeaderContainer'
 import TransactionFloaterContainer from 'containers/TransactionFloaterContainer'
 import EnableIntercom from 'containers/EnableIntercom'
-import { isConnectedToBlockchain, isConnectedToCorrectNetwork } from 'store/selectors/blockchain'
-import { getActiveProvider } from 'integrations/store/selectors'
+import { isConnectedToBlockchain } from 'store/selectors/blockchain'
+import { getActiveProvider, isConnectedToCorrectNetwork } from 'integrations/store/selectors'
 import { changeUiState } from 'store/actions/interface'
 import { getUiState } from 'store/selectors/interface'
